requirements.txt
certifi==2019.11.28
chardet==3.0.4
darkskyapi-py==1.1.0
idna==2.8
requests==2.22.0
urllib3==1.25.7
